requirements.txt
python_dotenv
requests
discord